requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
geographiclib==2.0
geopy==2.4.1
psycopg==3.1.18
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
